--- Java/Barcode/.idea/encodings.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="Encoding">
-    <file url="file://$PROJECT_DIR$/Barcode/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/Barcode/src/main/resources" charset="UTF-8" />
-  </component>
-</project>
