--- Java/Barcode/Barcode/src/main/java/org/example/BarcodeGenerator.java
@@ -1,13 +1,0 @@
-import javax.swing.SwingUtilities;
-
-public class BarcodeGenerator {
-
-    public static void main(String[] args) {
-
-        SwingUtilities.invokeLater(() -> {
-            BarcodeGeneratorGUI gui = new BarcodeGeneratorGUI();
-            gui.setVisible(true);
-        });
-        //just a random comment
-    }
-}
