+++ Java/Barcode/Barcode_03_01/Barcode_03_01/src/main/resources/META-INF/MANIFEST.MF
@@ -1,0 +1,3 @@
+Manifest-Version: 1.0
+Main-Class: BarcodeGenerator
+
